Quick Sketch of Interface Changes for Group Support
Adds a quick sketch out of the interface and code changes needed to integrate "Plugin Groups" into the lazy loader plugin, as the former plugin is now unmaintained.

This is specifically done in a backwards-compatible approach, where the default groups are automatically created with the current loading options. This also does not include any UI work around how to integrate into the options.